// src/types/SiteUsageSummary.ts
export type SiteUsageSummary = {
  ID?: string;
  name?: string;
  appName?: string;
  siteName?: string;
  siteCode?: string;
  totalResult?: number;
  totalExecResultPercent?: number;
  totalExecResult?: number;
  totalResultPercent?: number;
  vertical?: string;
  totalOutputPercent?: string;
  totalOutput?: string;
};